fixed total_earn_from_points in update_loyalty API

+++ Client/models.py
@@ -438,7 +438,7 @@ class ClientLoyaltyPoint(models.Model):
     points_redeemed = models.FloatField(default=0, verbose_name='Total Number Of Points Redeemed')
 
     for_every_points = models.PositiveIntegerField(default=0, null=True, blank=True, verbose_name='For every (this) points')
-    customer_will_get_amount = models.PositiveIntegerField(default=0, null=True, blank=True, verbose_name='a customer will get (this) amount')
+    customer_will_get_amount = models.FloatField(default=0, null=True, blank=True, verbose_name='a customer will get (this) amount')
 
 
     created_at = models.DateTimeField(auto_now_add=now, null=True)
